Some refactoring and program to interface. but, I don't know how to make added features to make interface. extends exist interface? Isn't it complicate? Now that, Some Test class has Strong coupled Generator class. And Some Generator class's static methods are called by other classed, such as Checker. This coupling is very bad I think. But I'm not sure have to move th Util class. Because these method are very conhensive with that BlockCodes.

+++ src/main/flask/type/BitPattern.java
@@ -57,9 +57,9 @@ public class BitPattern {
 		return bits.get(index);
 	}
 
-	public void set(int decimal, int length) {
+	public void set(long decimal, int length) {
 		bits.clear();
-		String binary = Integer.toBinaryString(decimal);
+		String binary = Long.toBinaryString(decimal);
 		for (int i = 0; i < length - binary.length(); i++)
 			append(new Bit(0));
 		for (int i = Math.abs(Math.min(length - binary.length(), 0)); i < binary.length(); i++)
